Move SecretsList to separate component
