+++ src/pages/namespace/Settings/index.tsx
@@ -1,42 +1,15 @@
-import {
-  Dialog,
-  DialogClose,
-  DialogContent,
-  DialogFooter,
-  DialogHeader,
-  DialogTitle,
-  DialogTrigger,
-} from "~/design/Dialog";
-import { FC, useState } from "react";
-import {
-  FolderTree,
-  GitCommit,
-  Hexagon,
-  Key,
-  Palette,
-  PlusCircle,
-  SquareAsterisk,
-  Trash,
-} from "lucide-react";
-import { SubmitHandler, useForm } from "react-hook-form";
-import { Table, TableBody, TableCell, TableRow } from "~/design/Table";
-import { Trans, useTranslation } from "react-i18next";
+import { FolderTree, GitCommit, Hexagon, Key, Palette } from "lucide-react";
 import { useApiActions, useApiKey } from "~/util/store/apiKey";
 import { useNamespace, useNamespaceActions } from "~/util/store/namespace";
 import { useTheme, useThemeActions } from "~/util/store/theme";
 
 import Button from "~/design/Button";
 import { Card } from "~/design/Card";
-import Input from "~/design/Input";
-import { SecretSchemaType } from "~/api/secrets/schema";
-import { Textarea } from "~/design/TextArea";
-import { useCreateSecret } from "~/api/secrets/mutate/createSecret";
-import { useDeleteSecret } from "~/api/secrets/mutate/deleteSecret";
+import { FC } from "react";
+import SecretsList from "./SecretsList";
 import { useListNamespaces } from "~/api/namespaces/query/get";
-import { useSecrets } from "~/api/secrets/query/get";
+import { useTranslation } from "react-i18next";
 import { useVersion } from "~/api/version";
-import { z } from "zod";
-import { zodResolver } from "@hookform/resolvers/zod";
 
 const SettingsPage: FC = () => {
   const apiKey = useApiKey();
@@ -47,188 +20,16 @@ const SettingsPage: FC = () => {
 
   const selectedNamespace = useNamespace();
 
-  const secrets = useSecrets();
   const { data: version, isLoading: isVersionLoading } = useVersion();
   const { data: namespaces, isLoading: isLoadingNamespaces } =
     useListNamespaces();
 
-  const [dialogOpen, setDialogOpen] = useState(false);
-  const [deleteSecret, setDeleteSecret] = useState<SecretSchemaType>();
-  const [createSecret, setCreateSecret] = useState(false);
-
-  const { mutate: deleteSecretMutation } = useDeleteSecret({
-    onSuccess: () => {
-      setDeleteSecret(undefined);
-      setDialogOpen(false);
-    },
-  });
-
-  const { mutate: createSecretMutation } = useCreateSecret({
-    onSuccess: () => {
-      setCreateSecret(false);
-      setDialogOpen(false);
-    },
-  });
-
-  type SecretFormInput = {
-    name: string;
-    value: string;
-  };
-
-  const onSubmit: SubmitHandler<SecretFormInput> = ({ name, value }) => {
-    createSecretMutation({
-      name,
-      value,
-    });
-  };
-
-  const { register, handleSubmit, reset } = useForm<SecretFormInput>({
-    resolver: zodResolver(
-      z.object({
-        name: z.string(),
-        value: z.string(),
-      })
-    ),
-  });
-
   const { t } = useTranslation();
-
-  const resetDialog = (isOpening: boolean) => {
-    if (!isOpening) {
-      setDeleteSecret(undefined);
-      setCreateSecret(false);
-      reset();
-    }
-    setDialogOpen(isOpening);
-  };
 
   return (
     <div className="flex flex-col space-y-6 p-10">
       <section>
-        <Dialog
-          open={dialogOpen}
-          onOpenChange={(isOpening) => resetDialog(isOpening)}
-        >
-          <div className="mb-3 flex flex-row justify-between">
-            <h3 className="flex items-center gap-x-2 font-bold text-gray-10 dark:text-gray-dark-10">
-              <SquareAsterisk className="h-5" />
-              {t("pages.settings.secrets.list.title")}
-            </h3>
-
-            <DialogTrigger
-              asChild
-              data-testid="secret-create"
-              onClick={() => setCreateSecret(true)}
-            >
-              <Button variant="ghost">
-                <PlusCircle />
-              </Button>
-            </DialogTrigger>
-          </div>
-
-          <Card>
-            <Table>
-              <TableBody>
-                {secrets.data?.secrets.results.map((secret, i) => (
-                  <TableRow key={i}>
-                    <TableCell>{secret.name}</TableCell>
-                    <TableCell className="w-0">
-                      <DialogTrigger
-                        asChild
-                        data-testid="secret-delete"
-                        onClick={() => setDeleteSecret(secret)}
-                      >
-                        <Button variant="ghost">
-                          <Trash />
-                        </Button>
-                      </DialogTrigger>
-                    </TableCell>
-                  </TableRow>
-                ))}
-              </TableBody>
-            </Table>
-            {deleteSecret && (
-              <DialogContent>
-                <DialogHeader>
-                  <DialogTitle>
-                    <Trash /> Delete
-                  </DialogTitle>
-                </DialogHeader>
-                <div className="my-3">
-                  <Trans
-                    i18nKey="pages.settings.secrets.delete.description"
-                    values={{ name: deleteSecret.name }}
-                  />
-                </div>
-                <DialogFooter>
-                  <DialogClose asChild>
-                    <Button variant="ghost">Cancel</Button>
-                  </DialogClose>
-                  <Button
-                    data-testid="secret-delete-confirm"
-                    onClick={() =>
-                      deleteSecretMutation({ secret: deleteSecret })
-                    }
-                    variant="destructive"
-                  >
-                    Delete
-                  </Button>
-                </DialogFooter>
-              </DialogContent>
-            )}
-            {createSecret && (
-              <DialogContent>
-                <form
-                  id="create-secret"
-                  onSubmit={handleSubmit(onSubmit)}
-                  className="flex flex-col space-y-5"
-                >
-                  <DialogHeader>
-                    <DialogTitle>
-                      <Trash /> Create
-                    </DialogTitle>
-                  </DialogHeader>
-
-                  <fieldset className="flex items-center gap-5">
-                    <label
-                      className="w-[150px] text-right text-[15px]"
-                      htmlFor="name"
-                    >
-                      Name
-                    </label>
-                    <Input
-                      data-testid="new-secret-name"
-                      id="name"
-                      placeholder="secret-name"
-                      {...register("name")}
-                    />
-                  </fieldset>
-
-                  <fieldset className="flex items-start gap-5">
-                    <Textarea
-                      className="h-96"
-                      data-testid="new-workflow-editor"
-                      {...register("value")}
-                    />
-                  </fieldset>
-
-                  <DialogFooter>
-                    <DialogClose asChild>
-                      <Button variant="ghost">Cancel</Button>
-                    </DialogClose>
-                    <Button
-                      data-testid="secret-create-submit"
-                      type="submit"
-                      variant="primary"
-                    >
-                      Create
-                    </Button>
-                  </DialogFooter>
-                </form>
-              </DialogContent>
-            )}
-          </Card>
-        </Dialog>
+        <SecretsList />
       </section>
 
       <section>
